created authorization headers for all api requests

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,13 +1,29 @@
 import axios from "axios";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
-import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useNavigate } from "react-router-dom";
 import { HiOutlineTrash } from "react-icons/hi";
+
+import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
+
 const WorkoutDetails = ({ workout }) => {
+  const { user } = useAuthContext();
   const { dispatch } = useWorkoutsContext();
+  const navigate = useNavigate();
 
   const handleClick = async (e) => {
+    if (!user) {
+      navigate("/login");
+      return;
+    }
+
     const response = await axios.delete(
-      `http://localhost:4000/api/workouts/${workout._id}`
+      `http://localhost:4000/api/workouts/${workout._id}`,
+      {
+        headers: {
+          Authorization: `Bearer ${user.token}`,
+        },
+      }
     );
 
     if (response.status === 200) {
@@ -30,7 +46,7 @@ const WorkoutDetails = ({ workout }) => {
         {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
       </p>
       <span className="trash" onClick={handleClick}>
-        <HiOutlineTrash color="red"/>
+        <HiOutlineTrash color="red" />
       </span>
     </div>
   );
